pyproject.toml
[project]
name = "drg-server"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.13.*"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.136.0",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic>=2.13.2",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.38",
]

[dependency-groups]
dev = ["ruff>=0.14.9", "stats-code>=0.1.5", "ty>=0.0.31"]

[tool.ruff]
line-length = 120
target-version = "py313"
include = ["server/**/*.py"]
exclude = ["_*.py"]
